Cytomod can now fully output modified genomes.
Added functionality to do so in a genome-wide fashion. The ability to output an entire modified genome still requires some significant optimizations (it currently uses excessive memory).

Fixed a regression with WIG tracks where most loci were miscalled as modified. This was due to Genomedata placing 0s instead of NaN for WIG track regions without data.

Added parameters to save output to a FASTA file and to restrict output to a specific genomic region.