Resolved a regression WRT retrieval over an input region segmenting itself as if genome-wide retrieval was requested.
Attempting to implement automatic track prioritization.

For now, this is envisioned as an ad hoc and contrived means of determining which epigenetic modification has precedence. This would be done by naively considering the resolution, and secondarily, the frequency of a given type of modification. This attempt is presently hindered by an apparent inability of genomedata to retrieve the intervals over which track data is defined (i.e. to determine which have single-base resolution).