Added a '-R' parameter to allow the retrieval of modified sequence data within a randomly selected genomic region, of an optionally specified length. The program attempts to ensure that the sequence returned from the invocation of this parameter is located wholly within a supercontig.
Added the '-E' parameter to allow for the use of only autosoms, only allosomes, or to exclude the mitochondrial chromosome.

Modularized region determination, conducted some general refactoring, and resolved some regressions which occasionally caused output to erroneously be directed to STDOUT instead of to a file. Now the '-f' parameter always causes output to a file instead of STDOUT. This has been clarified in the usage information.