Fixed a few regressions. One caused an erroneously modified sequence to result if not all bases in a sequence were modified. The others caused occasional index out of bounds errors or an err or if no bases were modified.
We now directly modify the reference sequence at the correct loci, in a vectorized fashion.

Again, this refactoring made the code slightly faster, 4.5 seconds for the same 1 000 000 random bases (chr19:33280172-34280172) as mentioned previously.

We are now ready to create BED tracks, since we now have easy access to NumPy arrays of both the indicies of modification and the modifications themselves.