baseFrequency.py
Added ability to transperantly handle gzipped FASTA file input.

Removed "sys.path.reverse()", as it is no longer needed on my system and it could cause issues on other systems.

Cytomod.py:

Changed BED track output to write headers before computing the modified genome in blocks. This allowed for existing files to be removed, so that we are not appedning to old data. It also resolved a regression wherein the track line would be appended to the file numerous times (and not just as the fist line of the file).