Adding generateMinimalMEMETextFromSeqs.py (whose docstring is reproduced below):
Creates a minimal MEME text file, suitable for FIMO, from an input set of sequences (one raw sequence per line). Accepts a single argument: the full path to the input file. The background frequencies should be manually adjusted to use case. NB: This script is not intended for general use, but rather was created prioritizing coding efficiency, for a specific use case. As such, it is not optimized towards ease-of-use.