Fixed a regression wherein no default filename was provided for writing a FASTA file when computing an entire modified genome. Also fixed a regression that caused BED track files to be erroneously removed if the current region (as opposed to all regions) had no data for that track. Both of these regressions were able to cause premature termination of the program.
Added an ID code to generated track files, corresponding to the filename of the FASTA file being output, if applcabl. regression wherein no default filename was provided for writing a FASTA file when computing an entire modified genome. Also fixed a regression that caused BED track files to be erroneously removed if the current region (as opposed to all regions) had no data for that track. Both of these regressions were able to cause premature termination of the program.

Added an ID code to generated track files, corresponding to the filename of the FASTA file being output, if applicable.