Added minor tick marks (4-5, automatically placed) for percent frequency plots.
Also added a function parameter for both plot functions (i.e. count and frequency) to annotate horizontal bar plots with the actual value plotted. This currently defaults to False and we do not currently have any plans to implement any means of activating it (it was mainly used for internal testing).