Fixed the underlying issue with list to string concatenation, reverting the change in the previous commit. This also resolves errors introduced by the previous commit.
The @_unpacked custom decorator used in cUtils.complement was incorrectly
defined, in that it checked the length of the output to determine if a
scalar or list ought to be returned. It should have instead checked if
the input was a scalar to make this determination. It now does so,
testing for any Iterable abstract base class, other than strings.

Amended function docstrings accordingly and to improve clarity.